+++ lib/runners/results.rb
@@ -113,30 +113,6 @@ module Runners
       end
     end
 
-    # Result to indicate failure because of absent of files.
-    # At least one of given files must be present for analysis.
-    # Client programs may return this result.
-    class MissingFilesFailure < Base
-      # @dynamic files
-      attr_reader :files
-
-      def initialize(guid:, files:)
-        super(guid: guid)
-        @files = files
-      end
-
-      def as_json
-        super.tap do |json|
-          json[:type] = "missing_files"
-          json[:files] = files.map(&:to_s)
-        end
-      end
-
-      def valid?
-        super && !files.empty?
-      end
-    end
-
     # Result to indicate that processor raises an exception.
     # Client programs should not return this result.
     class Error < Base